requirements.txt
cryptography==41.0.1
python-dotenv==1.0.0
onnx==1.14.0
onnxruntime==1.15.0
numpy>=1.24.0
joblib==1.3.1
scikit-learn==1.3.0